+++ modules/features/player/src/main/java/au/com/shiftyjelly/pocketcasts/player/view/bookmark/BookmarksSortByDialog.kt
@@ -3,10 +3,10 @@ package au.com.shiftyjelly.pocketcasts.player.view.bookmark
 import android.content.Context
 import androidx.fragment.app.FragmentManager
 import au.com.shiftyjelly.pocketcasts.analytics.SourceView
-import au.com.shiftyjelly.pocketcasts.models.type.BookmarksSortType
-import au.com.shiftyjelly.pocketcasts.models.type.BookmarksSortTypeForPlayer
-import au.com.shiftyjelly.pocketcasts.models.type.BookmarksSortTypeForPodcast
 import au.com.shiftyjelly.pocketcasts.preferences.Settings
+import au.com.shiftyjelly.pocketcasts.preferences.model.BookmarksSortType
+import au.com.shiftyjelly.pocketcasts.preferences.model.BookmarksSortTypeDefault
+import au.com.shiftyjelly.pocketcasts.preferences.model.BookmarksSortTypeForPodcast
 import au.com.shiftyjelly.pocketcasts.views.dialog.OptionsDialog
 import au.com.shiftyjelly.pocketcasts.localization.R as LR
 
@@ -17,27 +17,40 @@ class BookmarksSortByDialog(
     private val forceDarkTheme: Boolean = false,
 ) {
     fun show(context: Context, fragmentManager: FragmentManager) {
-        val message = "Bookmarks sort accessed in unknown source view: $sourceView"
-        val sortOrder = when (sourceView) {
-            SourceView.PLAYER -> settings.getBookmarksSortTypeForPlayer()
-            SourceView.PODCAST_SCREEN -> settings.getBookmarksSortTypeForPodcast()
-            else -> throw IllegalAccessError(message)
-        }
         val dialog = OptionsDialog()
             .setTitle(context.resources.getString(LR.string.sort_by))
             .setForceDarkTheme(forceDarkTheme)
-        val valuesToShow = when (sourceView) {
-            SourceView.PLAYER -> BookmarksSortTypeForPlayer.values().toList()
-            SourceView.PODCAST_SCREEN -> BookmarksSortTypeForPodcast.values().toList()
-            else -> throw IllegalAccessError(message)
-        }
+        val sortOrder = sourceView.mapToBookmarksSortType()
+        val valuesToShow = getValuesToShow()
         for (order in valuesToShow) {
             dialog.addCheckedOption(
                 titleId = order.labelId,
-                checked = order.labelId == sortOrder.labelId,
+                checked = order.key == sortOrder.key,
                 click = { changeSortOrder(order) }
             )
         }
         dialog.show(fragmentManager, "bookmarks_sort_dialog")
+    }
+
+    private fun SourceView.mapToBookmarksSortType(): BookmarksSortType = when (this) {
+        SourceView.FILES,
+        SourceView.EPISODE_DETAILS,
+        -> settings.episodeBookmarksSortType.flow.value
+
+        SourceView.PLAYER -> settings.playerBookmarksSortType.flow.value
+        SourceView.PODCAST_SCREEN -> settings.podcastBookmarksSortType.flow.value
+        else -> throw IllegalAccessError("$UNKNOWN_SOURCE_MESSAGE $this")
+    }
+
+    private fun getValuesToShow(): List<BookmarksSortType> = when (sourceView) {
+        SourceView.FILES,
+        SourceView.EPISODE_DETAILS,
+        SourceView.PLAYER -> BookmarksSortTypeDefault.values().toList()
+        SourceView.PODCAST_SCREEN -> BookmarksSortTypeForPodcast.values().toList()
+        else -> throw IllegalAccessError("$UNKNOWN_SOURCE_MESSAGE $sourceView")
+    }
+
+    companion object {
+        private const val UNKNOWN_SOURCE_MESSAGE = "Bookmarks sort accessed in unknown source view:"
     }
 }