+++ modules/features/podcasts/src/main/java/au/com/shiftyjelly/pocketcasts/podcasts/view/podcast/adapter/BookmarkUpsellViewHolder.kt
@@ -9,8 +9,6 @@ import androidx.recyclerview.widget.RecyclerView
 import au.com.shiftyjelly.pocketcasts.analytics.SourceView
 import au.com.shiftyjelly.pocketcasts.compose.AppTheme
 import au.com.shiftyjelly.pocketcasts.compose.theme
-import au.com.shiftyjelly.pocketcasts.featureflag.Feature
-import au.com.shiftyjelly.pocketcasts.featureflag.FeatureTier
 import au.com.shiftyjelly.pocketcasts.player.view.bookmark.components.MessageViewColors
 import au.com.shiftyjelly.pocketcasts.player.view.bookmark.components.UpsellView
 import au.com.shiftyjelly.pocketcasts.settings.onboarding.OnboardingFlow
@@ -18,6 +16,8 @@ import au.com.shiftyjelly.pocketcasts.settings.onboarding.OnboardingLauncher
 import au.com.shiftyjelly.pocketcasts.settings.onboarding.OnboardingUpgradeSource
 import au.com.shiftyjelly.pocketcasts.ui.theme.Theme
 import au.com.shiftyjelly.pocketcasts.utils.extensions.getActivity
+import au.com.shiftyjelly.pocketcasts.utils.featureflag.Feature
+import au.com.shiftyjelly.pocketcasts.utils.featureflag.FeatureTier
 
 class BookmarkUpsellViewHolder(
     private val composeView: ComposeView,
@@ -35,7 +35,8 @@ class BookmarkUpsellViewHolder(
                         val source = OnboardingUpgradeSource.BOOKMARKS
                         val onboardingFlow = OnboardingFlow.Upsell(
                             source = source,
-                            showPatronOnly = Feature.BOOKMARKS_ENABLED.tier == FeatureTier.Patron,
+                            showPatronOnly = Feature.BOOKMARKS_ENABLED.tier == FeatureTier.Patron ||
+                                Feature.BOOKMARKS_ENABLED.isCurrentlyExclusiveToPatron(),
                         )
                         OnboardingLauncher.openOnboardingFlow(context.getActivity(), onboardingFlow)
                     },