+++ modules/services/featureflag/src/main/java/au/com/shiftyjelly/pocketcasts/featureflag/FeatureFlagWrapper.kt
@@ -1,7 +1,0 @@
-package au.com.shiftyjelly.pocketcasts.featureflag
-
-import javax.inject.Inject
-
-open class FeatureFlagWrapper @Inject constructor() {
-    open fun isEnabled(feature: Feature) = FeatureFlag.isEnabled(feature)
-}