+++ modules/services/featureflag/src/main/java/au/com/shiftyjelly/pocketcasts/featureflag/FeatureWrapper.kt
@@ -1,7 +1,0 @@
-package au.com.shiftyjelly.pocketcasts.featureflag
-
-import javax.inject.Inject
-
-class FeatureWrapper @Inject constructor() {
-    fun isAvailable(feature: Feature, userTier: UserTier) = Feature.isAvailable(feature, userTier)
-}