+++ modules/services/utils/src/main/java/au/com/shiftyjelly/pocketcasts/utils/featureflag/FeatureWrapper.kt
@@ -1,0 +1,8 @@
+package au.com.shiftyjelly.pocketcasts.utils.featureflag
+
+import javax.inject.Inject
+
+class FeatureWrapper @Inject constructor() {
+    fun isAvailable(feature: Feature, userTier: UserTier) = Feature.isUserEntitled(feature, userTier)
+    val bookmarksFeature = Feature.BOOKMARKS_ENABLED
+}