+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 7.50
 -----
 
+*   Updates:
+    *   Show if a podcast has no ratings
+        ([#1453](https://github.com/Automattic/pocket-casts-android/pull/1453))
+*   Bug Fixes:
+    *   Fixed show notes loading issues
+        ([#1436](https://github.com/Automattic/pocket-casts-android/pull/1436))
 
 7.49
 -----